+++ src/cheat/GameUtil.h
@@ -90,9 +90,6 @@ public:
 	// aka playermove flags (FL_ONGROUND etc - FL_* macros)
 	int get_player_flags();
 
-	void render_circle_opengl(float cx, float cy, float radius, int num_segments, float width, bool blend, int r, int g, int b, int a);
-	void render_line_opengl(const Vector2D& from, const Vector2D& to, float width, bool blend, int r, int g, int b, int a);
-
 	// iuser3 cstrike flags
 	bool player_can_shoot();
 	bool player_freeze_time_over();